--- a/config/application.rb
+++ b/config/application.rb
@@ -41,5 +41,6 @@ module Scorpio
 
     # These are defined in `/lib/*.rb
     require "scorpio"
+    require "parallel_dhis2"
   end
 end