--- a/config/initializers/feature_flipper.rb
+++ b/config/initializers/feature_flipper.rb
@@ -1,0 +1,11 @@
+require 'flipper/adapters/active_record'
+
+Flipper.configure do |config|
+  config.default do
+    # pick an adapter, this uses memory, any will do
+    adapter = Flipper::Adapters::ActiveRecord.new
+
+    # pass adapter to handy DSL instance
+    Flipper.new(adapter)
+  end
+end