--- a/config/routes.rb
+++ b/config/routes.rb
@@ -10,6 +10,16 @@ Rails.application.routes.draw do
   Sidekiq::Throttled::Web.enhance_queues_tab!
   mount Sidekiq::Web => "/sidekiq"
 
+  if ENV["ADMIN_PASSWORD"]
+    flipper_app = Flipper::UI.app(Flipper) do |builder|
+      builder.use Rack::Auth::Basic do |username, password|
+        username == "admin" && password == ENV["ADMIN_PASSWORD"]
+      end
+    end
+  else
+    flipper_app = Flipper::UI.app(Flipper)
+  end
+  mount flipper_app => '/flipper'
   mount RailsAdmin::Engine => "/admin", as: "rails_admin"
 
   devise_for :users