… Gemfile.lock
… db/schema.rb
… spec/services/invoice_entity_spec.rb