## Why?

While publishing a rather large invoice, I found out that most of the time was spent in publishing to DHIS2.

When sending 11040 values to dhis2:

- Single request:    43 seconds (43093.83299993351)
- Parallel requests:  8 seconds (8223.681999836117)

Sending 3046 values had these results:

- Single request:  12 seconds
- Parallel rquests: 6 seconds

So big wins all around. I've also checked this on the DHIS2 side to confirm that the data is arriving.

``` sql      
select datavalue.lastupdated from datavalue 
  join organisationunit ON organisationunit.organisationunitid = datavalue.sourceid 
  join dataelement ON dataelement.dataelementid = datavalue.dataelementid 
where 
  dataelement.uid = 'WQbeZuIvhzS' and organisationunit.uid = 'XJkQtdGRO5Q';
```

And it was updating the lastupdated, it was also growing the audit table by the expected amounts.

## How?

`ParallelDhis2` uses the `typhoeus` gem to be able to make multiple concurrent requests to DHIS2, which is faster than the one huge request. ParallelDhis2 tries to mimick as much of its behavior from the normal `Dhis2::Client`, so that changes to that client or configuration are automatically taken over by ParallelDhis2.

### What's with the one huge file?

1. It's basically one class with two helper classes
2. It has a potential to be reused in other projects, so one file is
quicker to test out than a couple of files
3. I thought it was going to be smaller

## Production?

At the moment this feature is behind a feature flag (feature flags, are handled by `flipper` another gem I pulled in, to help with feature flags). The idea is to enable it for one specific `ProjectAnchor` so we can monitor the performance and results before enabling it by default.

### How to enable?

Either a console: `Flipper[:use_parallel_publishing].enable?(ProjectAnchor.find(<your-id_here>)`

Or visit `/flipper` in production and enable it using the UI (it's password protected).

## Aftermath

I've made a small adjustment to the output of the Dhis2Log that gets written by `ParallelDhis` runs, in the description will be `[parallel]` at the end. This is a small way to visually check who did the publishing.

The following snippet should paint a picture on whether it's working or not.

``` ruby
table = InvoicingJob.arel_table
anchor_id = <your_id_here>
before = InvoicingJob.where(project_anchor_id: anchor_id).where(table[:created_at].lt("20190214")).where.not(table[:duration_ms].eq(nil)).average(:duration_ms).to_f
after = InvoicingJob.where(project_anchor_id: anchor_id).where(table[:created_at].gt("20190214")).where.not(table[:duration_ms].eq(nil)).average(:duration_ms).to_f
puts "Before #{before} and After #{after}"
```

## Self proof reading checklist

- [x] Did I run Pronto and fixed all warnings (pronto run -c origin/dev)
- [x] Make sure all naming and code will remain understandable in 6 month by someone new to the project or by you.
- [x] Did I test the right thing?
- [x] Did I test corner cases, non happy path (defensive testing)?
- [x] Check that I used the ad-hoc patterns and created files accordingly (Presenters, Services, Search object, Form object,...)
- [x] Check code efficiency with record intensive project
- [x] Update documentation,readme accordingly

Thanks!