@pjaspers 2019-02-25
This turned out not be such a good idea:

```
table = InvoicingJob.arel_table
anchor_id = <your_id_here>
before = InvoicingJob.where(project_anchor_id: anchor_id).where(table[:created_at].lt("20190214")).where.not(table[:duration_ms].eq(nil)).average(:duration_ms).to_f
after = InvoicingJob.where(project_anchor_id: anchor_id).where(table[:created_at].gt("20190214")).where.not(table[:duration_ms].eq(nil)).average(:duration_ms).to_f
puts "Before #{before} and After #{after}"
```

The `InvoicingJob` gets reused, so the durations are overwritten, so this is not a good way to check if things are going better or worse. @mestachs would I be able to get it out of the DHIS2-logs?